The Level "buffint" should be short, not int, when forming the metadata.

+++ ogdi/driver/vrf/utils.c
@@ -17,7 +17,10 @@
  ******************************************************************************
  *
  * $Log$
- * Revision 1.14  2004-02-18 21:33:18  warmerda
+ * Revision 1.15  2004-10-25 19:34:31  warmerda
+ * The Level "buffint" should be short, not int, when forming the metadata.
+ *
+ * Revision 1.14  2004/02/18 21:33:18  warmerda
  * free regex memory
  *
  * Revision 1.13  2003/05/21 18:47:31  warmerda
@@ -551,7 +554,7 @@ vrf_GetMetadata(s)
   row_type rowcomp;
   row_type rowfca;
   float buffloat;
-  int buffint;
+  short buffint;
   register ServerPrivateData *spriv = s->priv;
   char buffer[256];
   char tab[3][7]={"char","float","int"};
